+++ tests/views.py
@@ -2,48 +2,23 @@
 
 from __future__ import unicode_literals
 import json
-import re
-import time
-from datetime import datetime
 
 from django.http import Http404, HttpResponse
 
 from infinite_scroll_pagination import paginator
+from infinite_scroll_pagination import serializers
 
 from .models import Article
-
-PAGE_RE = re.compile(r'^(?P<value>[0-9]+\.[0-9]+)-(?P<pk>[0-9]+)$')
-
-
-# XXX move to lib
-def page_key(raw_page):
-    if not raw_page:
-        return None, None
-    m = re.match(PAGE_RE, raw_page)
-    if m is None:
-        raise Http404()
-    try:
-        timestamp = datetime.fromtimestamp(float(m.group('value')))
-    except (OverflowError, ValueError):
-        raise Http404()
-    return timestamp, m.group('pk')
-
-
-def to_page_key(value=None, pk=None):
-    if value is None:
-        return ''
-    try:
-        timestamp = value.timestamp()
-    except AttributeError:
-        timestamp = time.mktime(value.timetuple())
-    return '{}-{}'.format(timestamp, pk)
 
 
 def pagination_ajax(request):
     if not request.is_ajax():
         return Http404()
 
-    value, pk = page_key(request.GET.get('p', ''))
+    try:
+        value, pk = serializers.page_key(request.GET.get('p', ''))
+    except serializers.InvalidPage:
+        return Http404()
 
     try:
         page = paginator.paginate(
@@ -64,7 +39,7 @@ def pagination_ajax(request):
             'prev_objects_left': page.prev_objects_left(limit=100),
             'next_pages_left': page.next_pages_left(limit=100),
             'prev_pages_left': page.prev_pages_left(limit=100),
-            'next_page': to_page_key(**page.next_page()),
-            'prev_page': to_page_key(**page.prev_page())}
+            'next_page': serializers.to_page_key(**page.next_page()),
+            'prev_page': serializers.to_page_key(**page.prev_page())}
 
     return HttpResponse(json.dumps(data), content_type="application/json")
